Refactored Encrypt class to be separate from Storage
