+++ match/lib/match/encryption/interface.rb
@@ -1,0 +1,17 @@
+module Match
+  module Encryption
+    class Interface
+      # Call this method to trigger the actual
+      # encryption
+      def encrypt_files
+        not_implemented(__method__)
+      end
+
+      # Call this method to trigger the actual
+      # decryption
+      def decrypt_files
+        not_implemented(__method__)
+      end
+    end
+  end
+end
